Fix s2msi-reader to read products with arbitrary JP2 tiling scheme
The previous implementation with S2TileOpImage supposed
that the JAI tiling could be aligned with the JP2 tiling scheme,
in particular, a call to S2TileOpImage.computeRect was generating
at most a single JP2 decompression.
This is not valid anymore since recent updates on the PDGS side.
Some products even have a JP2 tiling scheme not aligned with powers of two
(640x640 internal tiling for 20m bands, for example).

This commit makes use of the JP2TileOpImage class from the jp2-reader module.
JP2TileOpImage is dedicated to reading one internal tile of a JP2 file,
in contrast with the older S2TileOpImage which was designed to handle one
full JP2 file.
This is much more convenient to use like this, since the MosaicOp now triggers
by itself the decompression of the different individual internal JP2 tiles
to fulfill JAI requests.

The mosaicing of internal JP2 tiles is done at the Sentinel2OrthoProductReader level
coping with translations due to internal tiling AND due to multi-L1C tiles mosaicing.