Fix SIITBX-228: in some products angles are not loaded and sometimes black lines appear in angle bands.
The origin for the mosaic is not computed directly from the upper-left tile because sometimes it is not present in the product.
The granules are cropped individually before creating the mosaic to remove some black lines.
The black lines are produced by the translation and it is not always produced in the last column and row.
New functions for cropping have been added:
 cropBordersIfAreZero()
 cropBordersOutsideScene()