the processing is divided into pre- and postprocessing with two different operators.
apart from getting the per tile information to estimate the best shift, the processing is done in "postprocessing" (which should therefore be considered as processing)

bestoffset is an array holding the best estimation of the shift for 0 all pixels, 1 land, 2 water pixels.

rectangleTile stores the rectangle coordinates of the tiles from the stack - and is used to catch an error, if a tile is produced twice. This error occured, if the processing has been run with 4G - probably it does no longer occur with 10G RAM.

Flag coding for shifted cloud mask is added.

the processing includes the clustering approach as well, which can be switched between 3 modes.

mountain shadow computation is included (with some fixes and using the overall cloud path).

Determination of the cloudID of continuous cloud areas.

PotentialCloudShadowAreaIdentifier returns both positions of potential cloud shadow and their respective offsets along the cloud path.
cloudshadowflagger uses (per cloudID) positions of potential cloud shadow, offsets of potential cloud shadows, the bestOffset (statistic based on all pixels).

the shifted cloud mask is set as a new flag. the optimum offset based on all pixels is used.